chore: move new backend-specific secrets logic to grow package
