+++ core/src/commands/cloud/secrets/secrets-list.ts
@@ -17,7 +17,7 @@ import { StringsParameter } from "../../../cli/params.js"
 import { styles } from "../../../logger/styles.js"
 import type { SecretResult } from "./secret-helpers.js"
 import { makeSecretFromResponse } from "./secret-helpers.js"
-import { handleSecretsUnavailableInNewBackend } from "../../../cloud/secrets.js"
+import { handleSecretsUnavailableInNewBackend } from "../../../cloud/grow/secrets.js"
 
 export const secretsListOpts = {
   "filter-envs": new StringsParameter({
